Refactoring contract tests directory structure
We missed fixing the contract tests directory structure on our previous
commit. This one fixes it.

Pairing with @carolinaborim
