+++ test/consumer-contracts/contracts/telemetry_api/equipment/equipment/getAnEquipment.js
