+++ test/consumer-contracts/contracts/telemetry_api/equipment/trackingData/searchByCanVariables.js
@@ -1,6 +1,6 @@
 import { Contract, Joi } from 'consumer-contracts';
 import config from '../../../../contract_config';
-import trackingDataSearchHelper from '../../../../../../src/helper/trackingDataSearchHelper';
+import trackingDataSearchHelper from '../../../../../../src/helpers/trackingDataSearch';
 import helper from '../../../../contract_helper';
 
 const canVariables = 'ENGINE_HOURS';
